replace uint32_t with uint64_t

+++ trunk/src/bitVec.cpp
@@ -92,7 +92,7 @@ size_t BitVec::size() const {
 }
 
 void BitVec::print() const {
-  for (uint32_t i = 0; i < size_; ++i){
+  for (size_t i = 0; i < size_; ++i){
     if (getBit(i)) cout << "1";
     else           cout << "0";
     if ((i+1)%8 == 0){
